Try to measure loss

+++ lib/ngtcp2_bbrfrcst.c
@@ -974,7 +974,7 @@ static int bbr_check_inflight_too_high(ngtcp2_bbr2_cc *bbr,
 
 static int is_inflight_too_high(ngtcp2_bbr2_cc *bbr, ngtcp2_conn_stat *cstat, const ngtcp2_rs *rs) {
   ngtcp2_log_info(bbr->ccb.log, NGTCP2_LOG_EVENT_RCV,
-                    "bbr2 op pckt_loss loss=%" PRIu64, rs->lost / rs->tx_in_flight);
+                    "bbr2 op pckt_loss loss=%f", (float)rs->lost / rs->tx_in_flight);
   if (bbr->state == NGTCP2_BBRFRCST_STATE_FRCST) {
     return rs->lost * NGTCP2_BBR_LOSS_THRESH_DENOM >
           rs->tx_in_flight * cstat->frcst_loss;
